Update PcmPlayerEngine.kt
Linea 93 

Nueva::

val processors = Array(channels) { EqBank(sampleRate.toDouble()) }



Antigua::

val processors = Array(channels) { EqBank(sampleRate) }

+++ android/app/src/main/java/com/stylo/dsp/PcmPlayerEngine.kt
@@ -90,7 +90,7 @@ class PcmPlayerEngine(
                 .setTransferMode(AudioTrack.MODE_STREAM)
                 .build().also { it.setVolume(volume) }
 
-            val processors = Array(channels) { EqBank(sampleRate) }
+            val processors = Array(channels) { EqBank(sampleRate.toDouble()) }
             var appliedVersion = -1L
             var inputDone = false
             var outputDone = false
